# LocalAIAgent.API/openapi/LocalAIAgent.API.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "LocalAIAgent.API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/auth/passkeys": {
      "post": {
        "tags": [
          "Fido2"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialRegistrationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialRegistrationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialRegistrationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RegisteredPublicKeyCredential"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisteredPublicKeyCredential"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisteredPublicKeyCredential"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Fido2"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CredentialInfo"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CredentialInfo"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CredentialInfo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/passkeys/remove": {
      "post": {
        "tags": [
          "Fido2"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "format": "byte"
              }
            },
            "text/json": {
              "schema": {
                "type": "string",
                "format": "byte"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string",
                "format": "byte"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttestationResponse": {
        "required": [
          "attestationObject",
          "clientDataJSON",
          "transports"
        ],
        "type": "object",
        "properties": {
          "attestationObject": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "clientDataJSON": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "transports": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthenticatorTransport"
            }
          }
        },
        "additionalProperties": false
      },
...
      "AuthenticationExtensionsClientOutputs": {
        "type": "object",
        "properties": {
          "example.extension.bool": {
            "type": "boolean",
            "nullable": true
          },
          "appid": {
            "type": "boolean"
          },
          "exts": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "uvm": {
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            },
            "nullable": true
          },
          "credProps": {
            "$ref": "#/components/schemas/CredentialPropertiesOutput"
          },
          "prf": {
            "$ref": "#/components/schemas/AuthenticationExtensionsPRFOutputs"
          },
          "largeBlob": {
            "$ref": "#/components/schemas/AuthenticationExtensionsLargeBlobOutputs"
          },
          "credProtect": {
            "$ref": "#/components/schemas/CredentialProtectionPolicy"
          }
        },
        "additionalProperties": false
      },
...
      "AuthenticationExtensionsLargeBlobOutputs": {
        "type": "object",
        "properties": {
          "supported": {
            "type": "boolean"
          },
          "blob": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "written": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "AuthenticationExtensionsPRFOutputs": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "results": {
            "$ref": "#/components/schemas/AuthenticationExtensionsPRFValues"
          }
        },
        "additionalProperties": false
      },
      "AuthenticationExtensionsPRFValues": {
        "required": [
          "first"
        ],
        "type": "object",
        "properties": {
          "first": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "second": {
            "type": "string",
            "format": "byte",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "AuthenticatorAttestationRawResponse": {
        "required": [
          "clientExtensionResults",
          "id",
          "rawId",
          "response",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "minLength": 1,
            "type": "string"
          },
          "rawId": {
            "type": "string",
            "format": "byte"
          },
          "type": {
            "$ref": "#/components/schemas/PublicKeyCredentialType"
          },
          "response": {
            "$ref": "#/components/schemas/AttestationResponse"
          },
          "extensions": {
            "$ref": "#/components/schemas/AuthenticationExtensionsClientOutputs"
          },
          "clientExtensionResults": {
            "$ref": "#/components/schemas/AuthenticationExtensionsClientOutputs"
          }
        },
        "additionalProperties": false
      },
...
      "AuthenticatorTransport": {
        "enum": [
          "usb",
          "nfc",
          "ble",
          "smart-card",
          "hybrid",
          "internal"
        ],
        "type": "string"
      },
...
      "CredentialInfo": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "regDate": {
            "type": "string",
            "format": "date-time"
          },
          "aaGuid": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "CredentialPropertiesOutput": {
        "type": "object",
        "properties": {
          "rk": {
            "type": "boolean"
          },
          "authenticatorDisplayName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CredentialProtectionPolicy": {
        "enum": [
          "userVerificationOptional",
          "userVerificationOptionalWithCredentialIDList",
          "userVerificationRequired"
        ],
        "type": "string"
      },
      "CredentialRegistrationRequest": {
        "required": [
          "attestation",
          "credentialName"
        ],
        "type": "object",
        "properties": {
          "attestation": {
            "$ref": "#/components/schemas/AuthenticatorAttestationRawResponse"
          },
          "credentialName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Fido2User": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PublicKeyCredentialType": {
        "enum": [
          "public-key",
          "invalid"
        ],
        "type": "string"
      },
      "RegisteredPublicKeyCredential": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/PublicKeyCredentialType"
          },
          "id": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "publicKey": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "transports": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthenticatorTransport"
            },
            "nullable": true
          },
          "signCount": {
            "type": "integer",
            "format": "int32"
          },
          "isBackupEligible": {
            "type": "boolean"
          },
          "isBackedUp": {
            "type": "boolean"
          },
          "aaGuid": {
            "type": "string",
            "format": "uuid"
          },
          "user": {
            "$ref": "#/components/schemas/Fido2User"
          },
          "attestationFormat": {
            "type": "string",
            "nullable": true
          },
          "attestationObject": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "attestationClientDataJson": {
            "type": "string",
            "format": "byte",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}